Added new core functions and updated CLI commands to use them.
