+++ mindflow/app/app.py
@@ -5,14 +5,14 @@ from mindflow.input import Arguments, Command
 from mindflow.settings import Settings
 from mindflow.state import STATE
 
-from mindflow.commands.ask import ask
+from mindflow.core.ask import ask
 
 # from mindflow.commands.config import config
 from mindflow.cli.new_click_cli.commands.delete import delete
 from mindflow.cli.new_click_cli.commands.diff import diff
 from mindflow.cli.new_click_cli.commands.index import index
 from mindflow.cli.new_click_cli.commands.inspect import inspect
-from mindflow.commands.query import query
+from mindflow.core.query import query
 
 
 def trim_json(data: dict, keys: list) -> dict:
