+++ mindflow/cli/new_click_cli/commands/ask.py
@@ -4,8 +4,7 @@
 
 import click
 
-from mindflow.settings import Settings
-from mindflow.utils.response import handle_response_text
+from mindflow.core.ask import run_ask
 
 @click.command(help="")
 @click.argument("prompt", type=str)
@@ -13,7 +12,4 @@ def ask(prompt: str) -> str:
     """
     This function is used to generate a prompt and then use it as a prompt for GPT bot.
     """
-    settings = Settings()
-    # Prompt GPT through Mindflow API or locally
-    response: str = settings.mindflow_models.query.model([{"role": "system", "content": "You are a helpful virtual assistant responding to a users query using your general knowledge and the text provided below."}, {"role": "user", "content": prompt}])
-    handle_response_text(response)
+    run_ask(prompt)
