+++ mindflow/cli/new_click_cli/commands/delete.py
@@ -3,9 +3,7 @@
 """
 
 from typing import List
-from mindflow.db.controller import DATABASE_CONTROLLER
-from mindflow.db.objects.document import Document
-from mindflow.resolving.resolve import resolve_all
+from mindflow.core.delete import run_delete
 
 import click
 
@@ -15,7 +13,4 @@ def delete(document_paths: List[str]):
     """
     This function is used to delete your MindFlow index.
     """
-    Document.delete_bulk([
-        document_reference.path for document_reference in resolve_all(document_paths)
-    ])
-    DATABASE_CONTROLLER.databases.json.save_file()
+    run_delete(document_paths)
