+++ mindflow/cli/run.py
@@ -12,14 +12,14 @@ from mindflow.input import Arguments, Command
 from mindflow.cli.parser import get_parsed_cli_args
 from mindflow.settings import Settings
 
-from mindflow.commands.ask import AskArgs, ask
+from mindflow.core.ask import AskArgs, ask
 from mindflow.cli.new_click_cli.commands.commit import commit
 from mindflow.cli.new_click_cli.commands.config import config
 from mindflow.cli.new_click_cli.commands.delete import DeleteArgs, delete
 from mindflow.cli.new_click_cli.commands.diff import DiffArgs, diff
 from mindflow.cli.new_click_cli.commands.index import IndexArgs, index
 from mindflow.cli.new_click_cli.commands.inspect import InspectArgs, inspect
-from mindflow.commands.query import QueryArgs, query
+from mindflow.core.query import QueryArgs, query
 from mindflow.utils.response import handle_response_text
 
 MF_DESCRIPTION = """
